+++ core/src/main/java/org/patternfly/core/Dataset.java
@@ -17,8 +17,6 @@ package org.patternfly.core;
 
 public interface Dataset {
 
-    String expandableSectionId = "expandableSectionId";
-    String expandableSectionTarget = "expandableSectionTarget";
     String identifier = "identifier";
     String label = "label";
     String placeholder = "placeholder";